requirements.txt
# OpenAI Agents SDK (installs compatible pydantic and openai versions)
openai-agents>=0.1.0

# Model Context Protocol
mcp>=1.0.0

# HTTP Client
httpx>=0.27.0

# Notion API Client
notion-client>=2.2.1

# Environment & Configuration
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Logging
structlog>=24.1.0

# Scheduling
apscheduler>=3.10.0
pytz>=2024.1
croniter>=2.0.0

# Retry Logic
backoff>=2.2.0

# Database
aiosqlite>=0.20.0

# HTML Parsing
beautifulsoup4>=4.12.0
lxml>=5.2.0

# Markdown Parsing
mistletoe>=1.4.0
